Now testing obligations to revise
Hit a snag - the SyncAckPow isn't there after the agent SyncAck's (for
obligation to revise on syncack no) - also found 2 other obligation
issues. I think the "the agent had the power to do something when they
did it" problem is solveable by only inserting events that the agent
has the power to do into the kbase (so you never have to check for pow
in the kbase). This requires the refactoring of the powers and
permissions to generic forms. (!)

The other problem of saying "an agent was an acceptor before it left"
is possibly solvable by triggering on a flag event AgentLeftCluster(ag,
role, RIV) instead of on the agent action itself...