Potential solution to random infinite loop
It seemed that when LEaveCluster removed all the roles from an agent,
the leave/addrole pow/per rules were firing (as expected) but were also
somehow inserting the AddRole action fact, causing the agent to gain a
needtosync/sync fact, which the leavecluster rule kept trying to
remove, causing an infinite loop....

Added some of the LeaveCluster and AddRole rules to an agenda group set
to auto-focus with mostly lock-on-active set to true, so they can't
activate once one of the others has. Exception made for Per(AddRole)
which has to activate when Pow(AddRole) fires.