Partial implementation of agent-neutral obligations
Added fulfils() method to IPConAction - allows you to test if an action
(might) fulfil an action in an obligation (ie, if theyre the same
except for the agent). Each subclass implements it separately since
they have different fields.

Obligations now have extra fields : boolean agentNeutral and the role
that the obligation is for. If !agentNeutral, then role can be ignored.
Otherwise, agent should be null and everything else is handled
transparently in the query and in retraction.

Still need to sort out obligations when theyre is/was no leader at the
time. This means removing all the per/pow stuff from the obligations
(pulling in the relevant lines from the per/pow if necessary).