Think I've done all obligation-related shenanigens ! PossAddRev refactored into obligation insertion
PAR was being a pain and didn't retract when it was supposed to, so
refactored it into the obligation to revise on syncno. Will need to
remove it from the rest of the code. Not sure if PRR is suitable for
same treatment.

All agent-neutral obligations should now work properly.